addressed code review comments - just renaming/cleanup
