+++ cs/cs/cs.csproj
@@ -154,9 +154,12 @@
     <Compile Include="ObjectPool.cs" />
     <Compile Include="VowpalWabbitAsync.cs" />
     <Compile Include="VowpalWabbitDynamic.cs" />
+    <Compile Include="VowpalWabbitExampleCollection.cs" />
     <Compile Include="VowpalWabbitHttpModule.cs" />
     <Compile Include="VowpalWabbitJson.cs" />
     <Compile Include="VowpalWabbitJsonThreadedPrediction.cs" />
+    <Compile Include="VowpalWabbitMultiLineExampleCollection.cs" />
+    <Compile Include="VowpalWabbitSingleLineExampleCollection.cs" />
     <Compile Include="VowpalWabbitSweep.cs" />
     <Compile Include="VowpalWabbitThreadedPrediction.cs" />
     <Compile Include="VowpalWabbitThreadedLearning.cs" />
